change total amounts calculations to use long to avoid overflow

+++ app/src/main/java/com/romanoindustries/loanmanager/MainActivity.java
@@ -106,7 +106,7 @@ public class MainActivity extends AppCompatActivity {
         AlarmScheduler.scheduleAlarm(this);
     }
 
-    public static String formatAmount(int amount) {
+    public static String formatAmount(long amount) {
         return NumberFormat.getNumberInstance(Locale.US).format(amount).trim();
     }
 }
