+++ app/src/main/java/com/romanoindustries/loanmanager/fragments/IncomingLoansFragment.java
@@ -163,7 +163,7 @@ public class IncomingLoansFragment extends Fragment implements LoansAdapter.OnLo
         }
         SortModeHelper.sortLoansAccordingly(SortModeHelper.getSortMode(getContext()), loans);
         loansAdapter.updateLoans(loans);
-        int totalAmount = loans.stream().mapToInt(Loan::getCurrentAmount).sum();
+        long totalAmount = loans.stream().mapToLong(Loan::getCurrentAmount).sum();
         totalAmountTv.setText(MainActivity.formatAmount(totalAmount));
     }
